Add navigation, refactor main to home screen
